requirements.txt
ccxt
aiohttp
python-telegram-bot
